app/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncer==0.0.7
attrs==24.2.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.31.0
babel==2.16.0
backoff==2.2.1
bc-detect-secrets==1.5.15
bcrypt==4.2.0
bidict==0.23.1
blinker==1.8.2
blis==0.7.11
build==1.2.2.post1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
chainlit==1.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.6
chromadb==0.5.11
click==8.1.7
cloudpathlib==0.19.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
courlan==1.3.1
cymem==2.0.8
dataclasses-json==0.6.7
dateparser==1.2.0
debugpy==1.8.6
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
docx2txt==0.8
durationpy==0.9
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
Faker==27.4.0
fastapi==0.112.4
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
fuzzysearch==0.7.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
htmldate==1.9.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.28.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
json_repair==0.28.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jusText==3.0.1
kubernetes==31.0.0
langchain==0.3.2
langchain-community==0.3.1
langchain-core==0.3.9
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langcodes==3.4.1
langsmith==0.1.131
language_data==1.2.0
Lazify==0.4.0
literalai==0.0.607
llm-guard==0.3.15
lxml==5.3.0
lxml_html_clean==0.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
msrest==0.7.1
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.9.1
nest-asyncio==1.6.0
networkx==3.4
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
oldest-supported-numpy==0.9
onnxruntime==1.19.2
openai==1.51.0
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
phonenumbers==8.13.47
pillow==10.4.0
platformdirs==4.3.6
posthog==3.7.0
preshed==3.0.9
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
streamlit==1.39.0
structlog==24.4.0
sympy==1.13.3
syncer==2.0.3
tabulate==0.9.0
tenacity==8.5.0
thinc==8.2.5
tiktoken==0.7.0
tld==0.13
tldextract==5.1.2
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.2
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
trafilatura==1.12.2
traitlets==5.14.3
transformers==4.45.2
triton==3.0.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
unidiff==0.7.5
uptrace==1.27.0
urllib3==2.2.3
uvicorn==0.25.0
uvloop==0.20.0
wasabi==1.1.3
watchdog==5.0.3
watchfiles==0.20.0
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
wsproto==1.2.0
yarl==1.13.1
zipp==3.20.2
